refactor: Rename svelte components
